\name{discreteHawkes-package}
\alias{discreteHawkes-package}
\alias{discreteHawkes}
\docType{package}
\title{
\packageTitle{discreteHawkes}
}
\description{
\packageDescription{discreteHawkes}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{discreteHawkes}
\packageIndices{discreteHawkes}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{discreteHawkes}

Maintainer: \packageMaintainer{discreteHawkes}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
